% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divajs.R
\name{shinydivajs}
\alias{shinydivajs}
\title{a JavaScript book image viewer}
\usage{
shinydivajs(iipServerURL, objectData, imageDir, inGrid = FALSE,
  enableAutoTitle = TRUE, enableFullscreen = TRUE, enableLinkIcon = TRUE,
  width = NULL, height = NULL)
}
\arguments{
\item{iipServerURL}{The URL to the IIPImage installation}

\item{objectData}{URL to the JSON file that provides the object dimension data}

\item{imageDir}{The absolute path of the directory that contains the 
images on the IIP Image Server}

\item{inGrid}{enableAutoTitle}

\item{enableAutoTitle}{Shows the title within a div of id diva-title}

\item{enableFullscreen}{Enable or disable fullscreen icon (mode still available)}

\item{enableLinkIcon}{Controls the visibility of the link icon}

\item{width}{Width in pixels (optional, defaults to automatic sizing)}

\item{height}{Height in pixels (optional, defaults to automatic sizing)}
}
\value{
Interactive dygraph plot
}
\description{
R interface to a JavaScript book image viewer designed to 
present multi-page documents at multiple resolutions using 
\href{https://github.com/DDMAL/diva.js}{diva.js} JavaScript library.
}

